Renamed webpack config files and placed them all in config directory
